requirements.txt
crewai
langgraph
langchain-openai
pydantic
python-dotenv
